Fix: 상단 MEC 동작에러 해결
